+++ projects/ngx-libs/src/lib/ngx-libs.module.ts
@@ -1,10 +1,12 @@
 import { NgModule } from '@angular/core';
-import { NgxShadeModule } from './ngx-shade/ngx-shade.module';
-import { NgxPrismjsModule } from './ngx-prismjs/ngx-prismjs.module';
+import { NgxShadeModule } from './ngx-shade/index';
+import { NgxPrismjsModule } from './ngx-prismjs/index';
+import { NgxPerfectScrollbarModule } from './ngx-perfect-scrollbar/index';
 
 const MODULES = [
   NgxPrismjsModule,
   NgxShadeModule,
+  NgxPerfectScrollbarModule
 ];
 
 @NgModule({