+++ projects/ngx-libs/src/lib/ngx-perfect-scrollbar/classes/SCROLL_DIRECTION.enum.ts
@@ -1,0 +1,9 @@
+/** 滚动方向
+ *
+ *
+ * @export
+ */
+export enum SCROLL_DIRECTION {
+    SCROLL_TOP = 'scrollTop',
+    SCROLL_LEFT = 'scrollLeft',
+}