+++ projects/ngx-libs/src/lib/ngx-perfect-scrollbar/classes/index.ts
@@ -1,0 +1,3 @@
+export {
+    SCROLL_DIRECTION
+} from './SCROLL_DIRECTION.enum';